Move entities to their own files
