+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/domain/entities/CommentContent.kt
@@ -1,0 +1,4 @@
+package com.neaniesoft.vermilion.postdetails.domain.entities
+
+@JvmInline
+value class CommentContent(val value: String)
