+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/domain/entities/CommentFlags.kt
@@ -1,0 +1,10 @@
+package com.neaniesoft.vermilion.postdetails.domain.entities
+
+enum class CommentFlags {
+    SAVED,
+    EDITED,
+    IS_OP,
+    STICKIED,
+    SCORE_HIDDEN,
+    LOCKED
+}
