+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/domain/entities/CommentId.kt
@@ -1,0 +1,4 @@
+package com.neaniesoft.vermilion.postdetails.domain.entities
+
+@JvmInline
+value class CommentId(val value: String)
