+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/domain/entities/DurationString.kt
@@ -1,0 +1,4 @@
+package com.neaniesoft.vermilion.postdetails.domain.entities
+
+@JvmInline
+value class DurationString(val value: String)
