+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/domain/entities/UpVotesCount.kt
@@ -1,0 +1,4 @@
+package com.neaniesoft.vermilion.postdetails.domain.entities
+
+@JvmInline
+value class UpVotesCount(val value: Int)
